Add a new method named position()
This method returns the position of the next token. This might not seem
particularly difficult, but the lexer ignores whitespace. However, when
displaying error messages and reporting position, we can't ignore
whitespace.

The already existing API is untouched in terms of semantics, but
internally, tokens are stored together with their positions. This allows
us to report position while still ignoring whitespace. This feature will
eventually be used by `xpath-analyzer`.